Adjust help messaging, move sync to its own util
